Bringing in all changes from 0rtis' original up current with this commit: https://github.com/0rtis/dfk/commit/4c85ccc1e78f4d6fd1bb0915024af56ec4529382 Changes from the origin are: - all code under `dfk_lib` directory - all example code under dfk_lib/examples
