+++ dfk_lib/examples/auction_example.py
@@ -2,6 +2,7 @@ import logging
 import sys
 import dfk_lib.auctions.hero.sale_auctions as hero_sales
 import dfk_lib.auctions.hero.rent_auctions as hero_rental
+import dfk_lib.hero.utils.utils as hero_utils
 from dfk_lib.auctions.auction import Auction
 import dfk_lib.auctions.land as land_auction
 import dfk_lib.auctions.utils.utils as auction_utils
@@ -23,13 +24,22 @@ if __name__ == "__main__":
     for auction in auctions:
         logger.info(str(auction))
 
-    # sales.bid_hero(hero_id, ether2wei(100), prv_key, nonce, gas_price_gwei, 30, rpc_server, logger)
 
     logger.info("\n")
     logger.info("Hero rental auctions:")
     auctions = hero_rental.get_open_auctions(graphql, 0, 10)
     for auction in auctions:
         logger.info(str(auction))
+
+    # serendale hero auction
+    logger.info(hero_sales.get_auction(hero_sales.SERENDALE_CONTRACT_ADDRESS, 181373, 'https://api.harmony.one'))
+    # hero_sales.bid_hero(hero_sales.SERENDALE_CONTRACT_ADDRESS, 181373, hero_sales.ether2wei(100), 'prv key', 'nonce', 50, 30)
+
+
+    # crystalvale hero auction
+    cv_hero_auctions = Auction(hero_sales.CRYSTLAVALE_CONTRACT_ADDRESS, 'https://subnets.avax.network/defi-kingdoms/dfk-chain/rpc', logger)
+    logger.info(auction_utils.human_readable_auction(cv_hero_auctions.get_auction(hero_utils.sd2cv_cv_hero_id(250))))
+    # cv_hero_auctions.bid_hero(hero_utils.sd2cv_cv_hero_id(250), hero_sales.ether2wei(100), 'prv key', 'nonce', 50, 30)
 
 
     # land
@@ -43,7 +53,7 @@ if __name__ == "__main__":
         auction = auction_utils.human_readable_auction(land_auction.auctions(i))
         logger.info("Auction index " + str(i) + ": " + str(auction))
 
-    auction = land_auction.get_auction(48)
+    auction = land_auction.get_auction(17)  # must be an open auction
     auction = auction_utils.human_readable_auction(auction)
     logger.info(str(auction))
     logger.info(str(land_auction.is_on_auction(33)))
