+++ dfk_lib/examples/hero_example.py
@@ -2,7 +2,8 @@ import logging
 import json
 import sys
 import dfk_lib.hero.utils.utils as hero_utils
-import dfk_lib.hero.hero as heroes
+import dfk_lib.hero.hero_core as hero_core
+from dfk_lib.hero.hero import Hero
 
 
 if __name__ == "__main__":
@@ -27,11 +28,25 @@ if __name__ == "__main__":
         last_names = hero_utils.parse_names(f.read())
     logger.info("Hero last name loaded")
 
-    # transfer(1, 'private key of the owner', 'next nonce of owner account', 'receiver address', 200, rpc_server, hero_abi_json, logger)
+    serendale_heroes = Hero(hero_core.SERENDALE_CONTRACT_ADDRESS, 'https://api.harmony.one', logger)
 
-    for i in range(1, 100):
-        logger.info("Processing hero #"+str(i))
-        owner = heroes.get_owner(i, rpc_server)
-        hero = heroes.get_hero(i, rpc_server)
-        readable_hero = heroes.human_readable_hero(hero, male_first_names, female_first_names, last_names)
+
+    #serendale_heroes.transfer(1, 'private key of the owner', 'next nonce of owner account', 'receiver address', 45, 30)
+
+    for i in range(1, 10):
+        logger.info("Processing serendale hero #"+str(i))
+        owner = serendale_heroes.get_owner(i)
+        hero = serendale_heroes.get_hero(i)
+        readable_hero = serendale_heroes.human_readable_hero(hero, male_first_names, female_first_names, last_names)
+        logger.info(json.dumps(readable_hero, indent=4, sort_keys=False) + "\n Owned by " + owner)
+
+    crystalvale_heroes = Hero(hero_core.CRYSTALVALE_CONTRACT_ADDRESS,
+                              'https://subnets.avax.network/defi-kingdoms/dfk-chain/rpc', logger)
+
+    for i in range(1, 10):
+        logger.info("Processing crystalvale hero #" + str(i))
+        cv_hero_id = hero_utils.sd2cv_cv_hero_id(i)
+        owner = crystalvale_heroes.get_owner(cv_hero_id)
+        hero = crystalvale_heroes.get_hero(cv_hero_id)
+        readable_hero = crystalvale_heroes.human_readable_hero(hero, male_first_names, female_first_names, last_names)
         logger.info(json.dumps(readable_hero, indent=4, sort_keys=False) + "\n Owned by " + owner)
