+++ dfk_lib/examples/realm.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class Realm(Enum):
+    SERENDALE = 1
+    CRYSTALVALE = 2
